Refactor HekaTransform to HekaIO.readFiles()
